+++ noodles-csi/src/reader.rs
@@ -238,6 +238,13 @@ fn read_bins<R>(reader: &mut R, depth: u8) -> io::Result<(HashMap<usize, Bin>, O
 where
     R: Read,
 {
+    fn duplicate_bin_error(id: usize) -> io::Result<(HashMap<usize, Bin>, Option<Metadata>)> {
+        Err(io::Error::new(
+            io::ErrorKind::InvalidData,
+            format!("duplicate bin ID: {id}"),
+        ))
+    }
+
     let n_bin = reader.read_i32::<LittleEndian>().and_then(|n| {
         usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
     })?;
@@ -257,12 +264,18 @@ where
             .map(bgzf::VirtualPosition::from)?;
 
         if id == metadata_id {
-            metadata = read_metadata(reader).map(Some)?;
+            let m = read_metadata(reader)?;
+
+            if metadata.replace(m).is_some() {
+                return duplicate_bin_error(id);
+            }
         } else {
             let chunks = read_chunks(reader)?;
             let bin = Bin::new(loffset, chunks);
-            // TODO: Check for duplicates.
-            bins.insert(id, bin);
+
+            if bins.insert(id, bin).is_some() {
+                return duplicate_bin_error(id);
+            }
         }
     }
 
@@ -370,6 +383,87 @@ mod tests {
             read_magic(&mut reader),
             Err(ref e) if e.kind() == io::ErrorKind::InvalidData
         ));
+    }
+
+    #[test]
+    fn test_read_bins() -> io::Result<()> {
+        const DEPTH: u8 = 5;
+
+        let data = [
+            0x00, 0x00, 0x00, 0x00, // n_bin = 0
+        ];
+        let mut reader = &data[..];
+        let (actual_bins, actual_metadata) = read_bins(&mut reader, DEPTH)?;
+        assert!(actual_bins.is_empty());
+        assert!(actual_metadata.is_none());
+
+        let data = [
+            0x02, 0x00, 0x00, 0x00, // n_bin = 2
+            0x00, 0x00, 0x00, 0x00, // bins[0].id = 0
+            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[0].loffset = 0
+            0x00, 0x00, 0x00, 0x00, // bins[0].n_chunk = 0
+            0x4a, 0x92, 0x00, 0x00, // bins[1].id = 37450
+            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[1].loffset = 0
+            0x02, 0x00, 0x00, 0x00, // bins[1].n_chunk = 2
+            0x62, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[1].metadata.ref_beg = 610
+            0x3d, 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[1].metadata.ref_end = 1597
+            0x37, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[1].metadata.n_mapped = 55
+            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[1].metadata.n_unmapped = 0
+        ];
+        let mut reader = &data[..];
+        let (actual_bins, actual_metadata) = read_bins(&mut reader, DEPTH)?;
+        assert_eq!(actual_bins.len(), 1);
+        assert!(actual_bins.get(&0).is_some());
+        assert!(actual_metadata.is_some());
+
+        let data = [
+            0x01, 0x00, 0x00, 0x00, // n_bin = 1
+        ];
+        let mut reader = &data[..];
+        assert!(matches!(
+            read_bins(&mut reader, DEPTH),
+            Err(e) if e.kind() == io::ErrorKind::UnexpectedEof
+        ));
+
+        let data = [
+            0x02, 0x00, 0x00, 0x00, // n_bin = 2
+            0x00, 0x00, 0x00, 0x00, // bins[0].id = 0
+            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[0].loffset = 0
+            0x00, 0x00, 0x00, 0x00, // bins[0].n_chunk = 0
+            0x00, 0x00, 0x00, 0x00, // bins[1].id = 0
+            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[1].loffset = 0
+            0x00, 0x00, 0x00, 0x00, // bins[1].n_chunk = 0
+        ];
+        let mut reader = &data[..];
+        assert!(matches!(
+            read_bins(&mut reader, DEPTH),
+            Err(e) if e.kind() == io::ErrorKind::InvalidData
+        ));
+
+        let data = [
+            0x02, 0x00, 0x00, 0x00, // n_bin = 2
+            0x4a, 0x92, 0x00, 0x00, // bins[0].id = 37450
+            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[0].loffset = 0
+            0x02, 0x00, 0x00, 0x00, // bins[0].n_chunk = 2
+            0x62, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[0].metadata.ref_beg = 610
+            0x3d, 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[0].metadata.ref_end = 1597
+            0x37, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[0].metadata.n_mapped = 55
+            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[0].metadata.n_unmapped = 0
+            0x4a, 0x92, 0x00, 0x00, // bins[1].id = 37450
+            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[1].loffset = 0
+            0x02, 0x00, 0x00, 0x00, // bins[1].n_chunk = 2
+            0x62, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[1].metadata.ref_beg = 610
+            0x3d, 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[1].metadata.ref_end = 1597
+            0x37, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[1].metadata.n_mapped = 55
+            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // bins[1].metadata.n_unmapped = 0
+        ];
+        let mut reader = &data[..];
+        assert!(matches!(
+            read_bins(&mut reader, DEPTH),
+            Err(e) if e.kind() == io::ErrorKind::InvalidData
+        ));
+
+        Ok(())
     }
 
     #[test]